requirements.txt
asgiref==3.5.2
attrs==21.4.0
Django
djangorestframework==3.13.1
django-extensions
django-rest-knox==4.2.0
django-allauth==0.54.0 # newer versions removed email_address_exists
drf-spectacular==0.21.2
# pywatchman # this was added to the requirements but python can't find the package: ERROR: Could not find a version that satisfies the requirement pywatchman==1.4.1 (from versions: none)
# So lets leave it out and see if nothing breaks, it can be removed later
GDAL
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
django-cors-headers==3.13.0
django-bulk-update-or-create==0.3.0
